+++ ZF/ZFLua/zfsrc/ZFLua/ZFLuaState.h
@@ -10,42 +10,47 @@
 ZF_NAMESPACE_GLOBAL_BEGIN
 
 /**
- * @brief global native lua state currently used by #ZFLuaExecute,
- *   ensured not null
+ * @brief return lua state for current thread,
+ *   auto create and attach if no state attached to current thread
  *
  * ZFFramework can be used in any native lua state (by using #ZFLuaExecute),
- * but only if you have attached them\n
- * by default #ZFLuaState would have a builtin state that already setup properly,
- * while you can attach your own native lua state to ZFFramework,
- * by #ZFLuaStateAttach/#ZFLuaStateDetach\n
- * the builtin lua state can also be changed by #ZFLuaStateChange
+ * but only if you have attached them by #ZFLuaStateAttach\n
+ * note that, lua is not thread safe,
+ * ZFFramework maintains lua states for each native #ZFThread,
+ * these lua state is completely separate from each other,
+ * no global data or data sharing available,
+ * you must supply global methods in cpp to share data between threads\n
+ * \n
+ * by default, #ZFLuaState would automatically create a thread local lua state for current thread
+ * (create new if not exist, you may change it by #ZFLuaStateChange),
+ * you may still attach your own lua state for same thread by #ZFLuaStateAttach,
+ * and obtain by #ZFLuaStateList for current thread,
+ * or #ZFLuaStateListForAllThread for all thread
  */
 ZFMETHOD_FUNC_DECLARE_0(ZFLIB_ZFLua, void *, ZFLuaState)
 
-/**
- * @brief change builtin lua state
- *
- * this method can not undo, old one would be closed
- */
+/** @brief change builtin lua state for current thread, see #ZFLuaState */
 ZFMETHOD_FUNC_DECLARE_1(ZFLIB_ZFLua, void, ZFLuaStateChange,
                         ZFMP_IN(void *, L))
 
-/** @brief see #ZFLuaState */
-ZFMETHOD_FUNC_DECLARE_1(ZFLIB_ZFLua, void, ZFLuaStateListT,
-                        ZFMP_IN_OUT(ZFCoreArray<void *> &, ret))
-/** @brief see #ZFLuaState */
+/** @brief get lua state list for current thread, see #ZFLuaState */
 ZFMETHOD_FUNC_DECLARE_0(ZFLIB_ZFLua, ZFCoreArrayPOD<void *>, ZFLuaStateList)
 
-/** @brief see #ZFLuaState */
+/** @brief get lua state list for all thread, use with caution, see #ZFLuaState */
+ZFMETHOD_FUNC_DECLARE_2(ZFLIB_ZFLua, void, ZFLuaStateListForAllThread,
+                        ZFMP_OUT(ZFCoreArray<void *>, luaStateList),
+                        ZFMP_OUT(ZFCoreArray<ZFThread *>, threadList))
+
+/** @brief create new lua state, see #ZFLuaState */
 ZFMETHOD_FUNC_DECLARE_0(ZFLIB_ZFLua, void *, ZFLuaStateOpen)
-/** @brief see #ZFLuaState */
+/** @brief close lua state, see #ZFLuaState */
 ZFMETHOD_FUNC_DECLARE_1(ZFLIB_ZFLua, void, ZFLuaStateClose,
                         ZFMP_IN(void *, L))
 
-/** @brief see #ZFLuaState */
+/** @brief attach existing lua state to current thread, see #ZFLuaState */
 ZFMETHOD_FUNC_DECLARE_1(ZFLIB_ZFLua, void, ZFLuaStateAttach,
                         ZFMP_IN(void *, L))
-/** @brief see #ZFLuaState */
+/** @brief detach lua state from current thread, see #ZFLuaState */
 ZFMETHOD_FUNC_DECLARE_1(ZFLIB_ZFLua, void, ZFLuaStateDetach,
                         ZFMP_IN(void *, L))
 
@@ -54,6 +59,7 @@ ZF_NAMESPACE_BEGIN(ZFGlobalEvent)
  * @brief see #ZFObject::observerNotify
  *
  * notified when #ZFLuaStateAttach,
+ * called in the same thread of #ZFLuaStateAttach called,
  * param0 is a #v_ZFPtr to lua state
  */
 ZFOBSERVER_EVENT_GLOBAL(ZFLIB_ZFLua, LuaStateOnAttach)
@@ -61,6 +67,7 @@ ZFOBSERVER_EVENT_GLOBAL(ZFLIB_ZFLua, LuaStateOnAttach)
  * @brief see #ZFObject::observerNotify
  *
  * notified when #ZFLuaStateDetach,
+ * called in the same thread of #ZFLuaStateDetach called,
  * param0 is a #v_ZFPtr to lua state
  */
 ZFOBSERVER_EVENT_GLOBAL(ZFLIB_ZFLua, LuaStateOnDetach)